refactor(game): create hook folder
