+++ src/components/game/hooks/useLives.ts
